cut line file
