+++ srcs/expands/cut_line.h
@@ -1,0 +1,10 @@
+#ifndef CUT_LINE_H
+# define CUT_LINE_H
+
+#include <stdlib.h>
+#include "../basic_enums.h"
+#include "../../libft/libft.h"
+
+//#define TST_CUT_LINE
+
+#endif
