+++ srcs/expands/expands.h
@@ -4,5 +4,4 @@
 # include "../basic_enums.h"
 # include "../../libft/libft.h"
 
-#define TST_CUT_LINE
 #endif
